Factor out common AskCityState component.
